refactor(VendingMachine): getter 삭제 및 미사용 클래스 삭제
